+++ README.md
@@ -1,7 +1,7 @@
 Documentation
 =============
 
-* Readthedocs: https://xfluo.readthedocs.io/en/latest/index.html
+* Readthedocs: https://xrftomo.readthedocs.io/en/latest/index.html
 
 Features
 ========
@@ -27,12 +27,12 @@ Installation
 
 Install with: 
 
-    $ conda install -c fabriciosm xfluo 
+    $ conda install -c fabriciosm xrftomo 
 
 OR install from source along with all dependencies:
 
-    $ git clone https://github.com/FabricioSMarin/xfluo.git
-    $ cd xfluo
+    $ git clone https://github.com/FabricioSMarin/xrftomo.git
+    $ cd xrftomo
     $ conda install -c conda-forge dxchange
     $ conda install -c conda-forge tomopy
     $ conda install -c conda-forge pyqt
@@ -48,41 +48,41 @@ OR install from source along with all dependencies:
 
     $ python setup.py install
     $ cd bin 
-    $ python xfluo init
+    $ python xrftomo init
 
 in the prepared virtualenv or as root for system-wide installation.
 
 Usage
 =====
 
-You can run xfluo using a command line, for a full list of options use: 
+You can run xrftomo using a command line, for a full list of options use: 
 
-    $ bin/xfluo gui -h
+    $ bin/xrftomo gui -h
 
 or a GUI with:
 
-    $ bin/xfluo gui
-    $ or double-click on xfluo.sh
+    $ bin/xrftomo gui
+    $ or double-click on xrftomo.sh
 
 ```
-If your python installation is in a location different from #!/usr/bin/env python please edit the first line of the bin/xfluo file to match yours.
+If your python installation is in a location different from #!/usr/bin/env python please edit the first line of the bin/xrftomo file to match yours.
 ```
 
 You can also load configuration parameters from a configuration file called
-`xfluo.conf`. You can create a template with
+`xrftomo.conf`. You can create a template with
 
-    $ xfluo init
-    $ or double-click xfluo_init.sh
+    $ xrftomo init
+    $ or double-click xrftomo_init.sh
 
 Contribute
 ==========
 
-* Issue Tracker: https://github.com/tomography/xfluo/issues
-* Documentation: https://github.com/tomography/xfluo/tree/master/docs
-* Source Code: https://github.com/tomography/xfluo/tree/master/xfluo
+* Issue Tracker: https://github.com/tomography/xrftomo/issues
+* Documentation: https://github.com/tomography/xrftomo/tree/master/docs
+* Source Code: https://github.com/tomography/xrftomo/tree/master/xrftomo
 
 License
 =======
 
 The project is licensed under the 
-`BSD-3 <https://github.com/tomography/xfluo/blob/master/LICENSE.txt>`_ license.
+`BSD-3 <https://github.com/tomography/xrftomo/blob/master/LICENSE.txt>`_ license.
