+++ docs/source/demo.rst
@@ -1,1 +1,1 @@
-Examples========Here we describe what the examples are doing. You can cite with :cite:`reference:01`... toctree::   demo/docs.demo.example_01.. automodule:: xfluo   :members:   :undoc-members:   :show-inheritance: 
+Examples========Here we describe what the examples are doing. You can cite with :cite:`reference:01`... toctree::   demo/docs.demo.example_01.. automodule:: xrftomo   :members:   :undoc-members:   :show-inheritance:
