+++ xrftomo/widgets/image_process_controls.py
