+++ xrftomo/widgets/sinogram_controls.py
