+++ xrftomo/widgets/sinogram_view.py
